refactor(whispering): rename the surface to transformation-picker
Greenfield clean break on the internal names left over from the
clipboard-only window. Rename the route /transform-clipboard ->
/transformation-picker, the window file and label to transformationPicker,
and the Tauri capability (identifier + windows target) to match, so the
window keeps its permissions. Split operations/transformation-clipboard.ts
by surface: the picker opener moves to operations/transformation-picker.ts;
the file keeps only the clipboard quick-run. Names now match the product
vocabulary (the 'Open transformation picker' shortcut, the /transformations
editor) end to end. Generated capability schemas regenerate on the next
Tauri build.
